+++ ftclib/src/main/java/com/epra/epralib/ftclib/storage/logdata/MotorControllerData.java
@@ -3,4 +3,4 @@ package com.epra.epralib.ftclib.storage.logdata;
 /**A record that stores data from a MotorController.
  *<p></p>
  * Queer Coded by Striker-909.*/
-public record MotorControllerData(long time, double power, int position, int target, double velocity, double targetVelocity, double pidTOutput, double pidVOutput) {}
+public record MotorControllerData(long time, double power, double position, double target, double velocity, double targetVelocity, double pidTOutput, double pidVOutput) {}
